Crate metro_m0

Re-exports

pub use cortex_m_rt::entry;

Modules

ac
Analog Comparators

adc
Analog Digital Converter

atsamd21g18a
Peripheral access API for ATSAMD21G18A microcontrollers (generated using svd2rust v0.13.1)

clock
Configuring the system clock sources. You will typically need to create an instance of GenericClockController before you can set up most of the peripherals on the atsamd21 device. The other types in this module are used to enforce at compile time that the peripherals have been correctly configured.

dac
Digital Analog Converter

delay
Delays

dmac
Direct Memory Access Controller

dsu
Device Service Unit

eic
External Interrupt Controller

evsys
Event System Interface

gclk
Generic Clock Generator

gpio
Working with GPIO pins. The pins are associated with the PORT hardware. This module defines a split method on the PORT type that is used to safely reference the individual pin configuration. The IO pins can be switched into alternate function modes, which routes the pins to different peripherals depending on the mode for the pin. The pin configuration is reflected through the use of type states to make the interface (ideally, or at least practically) impossible to misuse.

hal
A Hardware Abstraction Layer (HAL) for embedded systems

hmatrix
HSB Matrix

i2s
Inter-IC Sound Interface

mashup
The nightly-only concat_idents! macro in the Rust standard library is notoriously underpowered in that its concatenated identifiers can only refer to existing items, they can never be used to define something new.

mtb
Cortex-M0+ Micro-Trace Buffer

nvmctrl
Non-Volatile Memory Controller

pac0
Peripheral Access Controller 0

pm
Power Manager

port
Port Module

prelude
Import the prelude to gain convenient access to helper traits

rtc
Real-Time Counter

sercom
Working with the SERCOM peripherals.

sercom0
Serial Communication Interface 0

sysctrl
System Control

tc3
Basic Timer Counter 3

tcc0
Timer Counter Control 0

time
Time units

timer
Working with timer counter hardware

usb
Universal Serial Bus

usb
USB Device support

wdt
Watchdog Timer

Macros

define_pins
This macro is a helper for defining a Pins type in a board support crate. This type is used to provide more meaningful aliases for the various GPIO pins for a given board.

interrupt
Macro to override a device specific interrupt handler

Structs

AC
Analog Comparators

ADC
Analog Digital Converter

CBP
Cache and branch predictor maintenance operations

CPUID
CPUID

CorePeripherals
Core peripherals

DAC
Digital Analog Converter

DCB
Debug Control Block

DMAC
Direct Memory Access Controller

DSU
Device Service Unit

DWT
Data Watchpoint and Trace unit

EIC
External Interrupt Controller

EVSYS
Event System Interface

FPB
Flash Patch and Breakpoint unit

FPU
Floating Point Unit

GCLK
Generic Clock Generator

HMATRIX
HSB Matrix

I2S
Inter-IC Sound Interface

ITM
Instrumentation Trace Macrocell

MPU
Memory Protection Unit

MTB
Cortex-M0+ Micro-Trace Buffer

NVIC
Nested Vector Interrupt Controller

NVMCTRL
Non-Volatile Memory Controller

PAC0
Peripheral Access Controller 0

PAC1
Peripheral Access Controller 1

PAC2
Peripheral Access Controller 2

PM
Power Manager

PORT
Port Module

PORT_IOBUS
Port Module (IOBUS)

Peripherals
All the peripherals

Pins
Maps the pins to their arduino names and the numbers printed on the board.

RTC
Real-Time Counter

SCB
System Control Block

SERCOM0
Serial Communication Interface 0

SERCOM1
Serial Communication Interface 1

SERCOM2
Serial Communication Interface 2

SERCOM3
Serial Communication Interface 3

SERCOM4
Serial Communication Interface 4

SERCOM5
Serial Communication Interface 5

SYSCTRL
System Control

SYST
SysTick: System Timer

TC3
Basic Timer Counter 3

TC4
Basic Timer Counter 4

TC5
Basic Timer Counter 5

TCC0
Timer Counter Control 0

TCC1
Timer Counter Control 1

TCC2
Timer Counter Control 2

TPIU
Trace Port Interface Unit

USB
Universal Serial Bus

WDT
Watchdog Timer

Enums

Interrupt
Enumeration of all the interrupts

Constants

NVIC_PRIO_BITS
Number available in the NVIC for configuring priority

Functions

flash_spi_master
Convenience for accessing the on-board SPI Flash device. This powers up SERCOM5 and configures it for use as an SPI Master.

i2c_master
Convenience for setting up the labelled SDA, SCL pins to operate as an I2C master running at the specified frequency.

spi_master
Convenience for setting up the 2x3 header block for SPI. This powers up SERCOM4 and configures it for use as an SPI Master in SPI Mode 0. Unlike the flash_spi_master function, this one does not accept a CS pin; configuring a pin for CS is the responsibility of the caller, because it could be any OutputPin, or even a pulled up line on the slave.