Module metro_m0::atsamd21g18a::tcc0::wave

Waveform Control

Structs

CICCEN0R
Value of the field

CICCEN1R
Value of the field

CICCEN2R
Value of the field

CICCEN3R
Value of the field

CIPERENR
Value of the field

POL0R
Value of the field

POL1R
Value of the field

POL2R
Value of the field

POL3R
Value of the field

R
Value read from the register

SWAP0R
Value of the field

SWAP1R
Value of the field

SWAP2R
Value of the field

SWAP3R
Value of the field

W
Value to write to the register

_CICCEN0W
Proxy

_CICCEN1W
Proxy

_CICCEN2W
Proxy

_CICCEN3W
Proxy

_CIPERENW
Proxy

_POL0W
Proxy

_POL1W
Proxy

_POL2W
Proxy

_POL3W
Proxy

_RAMPW
Proxy

_SWAP0W
Proxy

_SWAP1W
Proxy

_SWAP2W
Proxy

_SWAP3W
Proxy

_WAVEGENW
Proxy

Enums

RAMPR
Possible values of the field RAMP

RAMPW
Values that can be written to the field RAMP

WAVEGENR
Possible values of the field WAVEGEN

WAVEGENW
Values that can be written to the field WAVEGEN