Trait metro_m0::gpio::IntoFunction

pub trait IntoFunction<T> {
    fn into_function(self, port: &mut Port) -> T;
}

A trait that makes it easier to generically manage converting a pin from its current state into some other functional mode. The configuration change requires exclusive access to the Port hardware, which is why this isn't simply the standard Into trait.

Required methods

fn into_function(self, port: &mut Port) -> T

Consume the pin and configure it to operate in the mode T.

Loading content...