Enum metro_m0::nvmctrl::ctrla::CMDR

pub enum CMDR {
    ER,
    WP,
    EAR,
    WAP,
    SF,
    WL,
    LR,
    UR,
    SPRM,
    CPRM,
    PBC,
    SSB,
    INVALL,
    _Reserved(u8),
}

Possible values of the field CMD

Variants
ER

Erase Row - Erases the row addressed by the ADDR register.

WP

Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register.

EAR

Erase Auxiliary Row - Erases the auxiliary row addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

WAP

Write Auxiliary Page - Writes the contents of the page buffer to the page addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

SF

Security Flow Command

WL

Write lockbits

LR

Lock Region - Locks the region containing the address location in the ADDR register.

UR

Unlock Region - Unlocks the region containing the address location in the ADDR register.

SPRM

Sets the power reduction mode.

CPRM

Clears the power reduction mode.

PBC

Page Buffer Clear - Clears the page buffer.

SSB

Set Security Bit - Sets the security bit by writing 0x00 to the first byte in the lockbit row.

INVALL

Invalidates all cache lines.

_Reserved(u8)

Reserved

Methods

impl CMDR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_er(&self) -> bool

Checks if the value of the field is ER

pub fn is_wp(&self) -> bool

Checks if the value of the field is WP

pub fn is_ear(&self) -> bool

Checks if the value of the field is EAR

pub fn is_wap(&self) -> bool

Checks if the value of the field is WAP

pub fn is_sf(&self) -> bool

Checks if the value of the field is SF

pub fn is_wl(&self) -> bool

Checks if the value of the field is WL

pub fn is_lr(&self) -> bool

Checks if the value of the field is LR

pub fn is_ur(&self) -> bool

Checks if the value of the field is UR

pub fn is_sprm(&self) -> bool

Checks if the value of the field is SPRM

pub fn is_cprm(&self) -> bool

Checks if the value of the field is CPRM

pub fn is_pbc(&self) -> bool

Checks if the value of the field is PBC

pub fn is_ssb(&self) -> bool

Checks if the value of the field is SSB

pub fn is_invall(&self) -> bool

Checks if the value of the field is INVALL