Module metro_m0::tcc0::evctrl

Event Control

Structs

CNTEOR
Value of the field

MCEI0R
Value of the field

MCEI1R
Value of the field

MCEI2R
Value of the field

MCEI3R
Value of the field

MCEO0R
Value of the field

MCEO1R
Value of the field

MCEO2R
Value of the field

MCEO3R
Value of the field

OVFEOR
Value of the field

R
Value read from the register

TCEI0R
Value of the field

TCEI1R
Value of the field

TCINV0R
Value of the field

TCINV1R
Value of the field

TRGEOR
Value of the field

W
Value to write to the register

_CNTEOW
Proxy

_CNTSELW
Proxy

_EVACT0W
Proxy

_EVACT1W
Proxy

_MCEI0W
Proxy

_MCEI1W
Proxy

_MCEI2W
Proxy

_MCEI3W
Proxy

_MCEO0W
Proxy

_MCEO1W
Proxy

_MCEO2W
Proxy

_MCEO3W
Proxy

_OVFEOW
Proxy

_TCEI0W
Proxy

_TCEI1W
Proxy

_TCINV0W
Proxy

_TCINV1W
Proxy

_TRGEOW
Proxy

Enums

CNTSELR
Possible values of the field CNTSEL

CNTSELW
Values that can be written to the field CNTSEL

EVACT0R
Possible values of the field EVACT0

EVACT0W
Values that can be written to the field EVACT0

EVACT1R
Possible values of the field EVACT1

EVACT1W
Values that can be written to the field EVACT1